pyproject.toml
[project]
name = "petri"
version = "0.1.0"
description = "Project Petri: A simulation platform with AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "litellm>=1.37.0",
    "chromadb>=1.5.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "openai>=2.24.0",
]
